+++ src/core/utils.h
@@ -1,0 +1,4 @@
+#pragma once
+#ifndef UTILS_H
+#define UTILS_H
+#endif // !UTILS_H