+++ src/levels/menu.c
@@ -6,23 +6,27 @@ void Menu_Run(Game* game, GameLevel* level, ecs_entity_t entity)
 	{
 
 		LoadSprite(game, "assets/window_icon.png", "test");
-		Sprite* ts = GetSprite(&game->resourcesManager, "test");
-		if (ts != NULL)
-			printf("Ancho: %d | Alto: %d\n", ts->width, ts->height);
 
 		LoadTilsetsPack(game, "assets/tilesets/pack_0.tspack", "base");
-		TilesetPack* pack = GetTilesetPack(&game->resourcesManager, "base");
-		if (pack != NULL)
-		{
-			printf("TILESET: %d\n", pack->tilesetsCount);
-			printf("El paquete se encuentra\n");
-		}
-
 		LoadTileMap(game, "assets/maps/farm_ext.tmx", "base", "test");
-		//TileMap* map = GetTileMap(&game->resourcesManager, "test");
-		//if (map != NULL)
-		//	printf("Mapa cargado\n");
+		LoadTileMap(game, "assets/maps/farmhouse.tmx", "base", "player_house");
 
 		level->isComponentsAdded = 1;
+	}
+
+	TileMap* map = GetTileMap(&game->resourcesManager, "player_house");
+	if (map != NULL)
+	{
+		DrawTexturePro(map->mapTexture.texture,
+			(Rectangle) {
+			0, 0, map->mapTexture.texture.width, -map->mapTexture.texture.height
+		},
+			(Rectangle) {
+			20, 20, map->mapTexture.texture.width * 4.0f, map->mapTexture.texture.height * 4.0f
+		},
+			(Vector2) {
+			0.0f, 0.0f
+		}, 0.0f, WHITE
+		);
 	}
 }